requirements.txt
# Core dependencies
MetaTrader5>=5.0.33
pymongo>=4.3.3
pandas>=1.3.0
numpy>=1.20.0

# API and web server
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6

# Technical analysis
ta>=0.7.0
matplotlib>=3.4.0

# Utilities
python-dotenv>=0.19.0
pytz>=2021.1
requests>=2.25.1

# Optional integrations
python-telegram-bot>=13.7
openai>=0.27.0
